foodcoop-api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1 